requirements.txt
ultralytics==8.3.22
opencv-python==4.9.0.80
scikit-learn==1.4.2
tqdm
pyyaml
pillow
geopy
scikit-image
